chore: refactored tests dir structure
