+++ packages/daemon/__tests__/integration/scripts/sequelize-config.js
