+++ packages/daemon/__tests__/integration/scripts/sequelize-db-config.js
